+++ src/main/java/com/dljsxy/school/model/CourseModel.java
@@ -4,8 +4,8 @@ import lombok.Data;
 @Data
 public class CourseModel {
     private String name;
-    private long credit;
-    private long teacher_id;
+    private short credit;
+    private int teacher_id;
 
 
 }
